Move an attribute & methods of TilesetTree into parent class
Move the leaf_nodes attribute and the getter methods from TilsetTree to GeometryTree
