+++ py3dtilers/TilesetReader/tileset_tree.py
@@ -1,4 +1,4 @@
-from ..Common import GeometryNode, GeometryTree, ObjectsToTile
+from ..Common import GeometryNode, GeometryTree
 
 
 class TilesetTree(GeometryTree):
@@ -48,27 +48,3 @@ class TilesetTree(GeometryTree):
             self.leaf_nodes.append(node)
 
         return node
-
-    def get_root_objects(self):
-        """
-        Return the geometries of the root nodes.
-        :return: list of geometries
-        """
-        return ObjectsToTile([node.objects_to_tile for node in self.root_nodes])
-
-    def get_leaf_objects(self):
-        """
-        Return the geometries of the leaf nodes.
-        :return: list of geometries
-        """
-        return ObjectsToTile([node.objects_to_tile for node in self.leaf_nodes])
-
-    def get_all_objects(self):
-        """
-        Return the geometries of all the nodes.
-        :return: list of geometries
-        """
-        objects = list()
-        for node in self.root_nodes:
-            objects.extend(node.get_objects())
-        return ObjectsToTile(objects)
